vm-configurator: Default to ich9 sound model
These days users can set which OS they are installing in case
libosinfo can't detect a given ISO. This way, it is very rare
when an user will perform an installation in which we can't figure
what's a adequate sound device model for it. And for those corner
cases, it is safe to assume that ich9 would likely work. If not,
the device sound model can be changed manually in the configuration.

This fixes issue #618, in which we were returning nothing when the
method signature had a (non-null) return type specified.

Fixes #618